refactor(choose-tree): extract overlay controller into useChooseTreeController
ChooseTreeOverlay had grown to ~120 lines of mixed imperative wiring
(refs, effects, callbacks, reducer setup, derived counts) and JSX. The
JSX kept getting harder to scan because every render had to mentally
skip past the controller plumbing first.

Move the imperative side to a new useChooseTreeController hook that
returns exactly what the JSX needs: { treeState, rows, cursor,
activeRowId, sessionCount, windowCount, rootRef, onBackdropPointerDown,
onRowClick }. The overlay shrinks to layout + KeyHint, no useState /
useReducer / useEffect / useRef at the top level.

Behaviour unchanged. 432/432 vitest pass, tsc/biome clean, production
build clean.
